requirements.txt
flask
google-cloud-storage
torch===1.5.0
torchvision===0.6.0
pandas